+++ wfd-xml/api/src/main/java/com/quadient/wfdxml/api/layoutnodes/TableStyle.java
@@ -1,0 +1,5 @@
+package com.quadient.wfdxml.api.layoutnodes;
+
+import com.quadient.wfdxml.api.Node;
+
+public interface TableStyle extends Node<TableStyle> { }