+++ wfd-xml/api/src/main/java/com/quadient/wfdxml/api/layoutnodes/tables/Table.java
@@ -57,6 +57,8 @@ public interface Table extends Node<Table> {
 
     Table setTablePdfAlternateText(String alternateText);
 
+    Table setExistingTableStyle(String id);
+
     enum TablePdfTaggingRule {
         NONE,
         DEFAULT,